create SpeedMath class for refactoring
