+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
+//    id 'kotlin-kapt'
 }
 
 android {
@@ -60,4 +61,28 @@ dependencies {
     // Room Database Dependencies
     implementation("androidx.room:room-runtime:$room_version")
     annotationProcessor("androidx.room:room-compiler:$room_version")
+
+    // To use Kotlin annotation processing tool (kapt)
+//    kapt("androidx.room:room-compiler:$room_version")
+    // To use Kotlin Symbol Processing (KSP)
+//    ksp("androidx.room:room-compiler:$room_version")
+
+//    // optional - Kotlin Extensions and Coroutines support for Room
+//    implementation("androidx.room:room-ktx:$room_version")
+//
+//    // optional - RxJava2 support for Room
+//    implementation("androidx.room:room-rxjava2:$room_version")
+//
+//    // optional - RxJava3 support for Room
+//    implementation("androidx.room:room-rxjava3:$room_version")
+//
+//    // optional - Guava support for Room, including Optional and ListenableFuture
+//    implementation("androidx.room:room-guava:$room_version")
+//
+//    // optional - Test helpers
+//    testImplementation("androidx.room:room-testing:$room_version")
+//
+//    // optional - Paging 3 Integration
+//    implementation("androidx.room:room-paging:$room_version")
+
 }
