+++ app/src/main/java/com/example/maths/GameOverFragment.kt
@@ -7,14 +7,15 @@ import android.view.ViewGroup
 import androidx.core.view.isInvisible
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.example.maths.MainActivity.Companion.bestTimeEasy
-import com.example.maths.MainActivity.Companion.bestTimeEasyString
-import com.example.maths.MainActivity.Companion.bestTimeHard
-import com.example.maths.MainActivity.Companion.bestTimeHardString
+import com.example.maths.SpeedMath.Companion.bestTimeEasy
+import com.example.maths.SpeedMath.Companion.bestTimeEasyString
+import com.example.maths.SpeedMath.Companion.bestTimeHard
+import com.example.maths.SpeedMath.Companion.bestTimeHardString
 import com.example.maths.databinding.FragmentGameOverBinding
-import com.example.maths.MainActivity.Companion.currScore
-import com.example.maths.MainActivity.Companion.gameTime
-import com.example.maths.MainActivity.Companion.formatTime
+import com.example.maths.SpeedMath.Companion.currScore
+import com.example.maths.SpeedMath.Companion.gameTime
+import com.example.maths.SpeedMath.Companion.formatTime
+import com.example.maths.SpeedMath.Companion.difficulty
 
 class GameOverFragment: Fragment() {
     private var _binding: FragmentGameOverBinding? = null
@@ -40,7 +41,7 @@ class GameOverFragment: Fragment() {
 
             binding.highScore.text = gameTimeFormatted
 
-            if (MainActivity.difficulty == 0) {
+            if (SpeedMath.difficulty == 0) {
                 binding.titleText.text = getString(R.string.easy_complete)
                 bestTimeEasyString = gameTimeFormatted
             } else {
@@ -56,9 +57,9 @@ class GameOverFragment: Fragment() {
     }
 
     private fun saveTimes() {
-        if (MainActivity.difficulty == 1 && (gameTime < bestTimeHard || bestTimeHard.compareTo(0) == 0)) {
+        if (difficulty == 1 && (gameTime < bestTimeHard || bestTimeHard.compareTo(0) == 0)) {
             bestTimeHard = gameTime
-        } else if (MainActivity.difficulty == 0 && (gameTime < bestTimeEasy || bestTimeEasy.compareTo(0) == 0)) {
+        } else if (difficulty == 0 && (gameTime < bestTimeEasy || bestTimeEasy.compareTo(0) == 0)) {
             bestTimeEasy = gameTime
         }
         val filename = "highScore"
