+++ app/src/main/java/com/example/maths/MainActivity.kt
@@ -15,44 +15,18 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
+import com.example.maths.SpeedMath.Companion.bestTimeEasy
+import com.example.maths.SpeedMath.Companion.bestTimeEasyString
+import com.example.maths.SpeedMath.Companion.bestTimeHard
+import com.example.maths.SpeedMath.Companion.bestTimeHardString
+import com.example.maths.SpeedMath.Companion.difficulty
+import com.example.maths.SpeedMath.Companion.gameMode
 import com.example.maths.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var appBarConfiguration: AppBarConfiguration
     private lateinit var binding: ActivityMainBinding
-
-    companion object {
-        var bestTimeHard: Long = 0
-        var bestTimeHardString: String = ""
-        var bestTimeEasy: Long = 0
-        var bestTimeEasyString: String = ""
-        var currScore = 0
-        var difficulty = 0
-        var gameMode = 0
-        var gameTime: Long = 0
-
-        fun formatTime(ms: Long): String {
-            val minutes = ms / 1000 /60
-            val seconds = ms / 1000 % 60
-
-            // Calculate decimal part of seconds
-            val doubleMinutesRaw: Double = ms.toDouble() / 1000
-            val doubleMinutesString = doubleMinutesRaw.toString()
-            val decimalStringHard = doubleMinutesString.substring(doubleMinutesString.indexOf("."))
-
-            val stringMinutes = minutes.toString()
-
-            val stringSeconds = if (seconds < 10) {
-                "0".plus(seconds.toString())
-            } else {
-                seconds.toString()
-            }
-            return stringMinutes.plus(":").plus(stringSeconds).plus(decimalStringHard)
-        }
-
-        var tts: TextToSpeech? = null
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -134,11 +108,11 @@ class MainActivity : AppCompatActivity() {
         // Do nothing
     }
 
-    override fun onDestroy() {
-        if (tts != null) {
-            tts!!.stop()
-            tts!!.shutdown()
-        }
-        super.onDestroy()
-    }
+//    override fun onDestroy() {
+//        if (tts != null) {
+//            tts!!.stop()
+//            tts!!.shutdown()
+//        }
+//        super.onDestroy()
+//    }
 }
