+++ build.gradle
@@ -1,6 +1,9 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.8.20'
+    ext {
+        kotlin_version = '1.8.20'
+        room_version = "2.5.2"
+    }
     repositories {
         google()
         mavenCentral()
